Split shell snapshot artifact tests
